// src/api/generated/models/MenuUpdateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MenuUpdateRequest
 */
export interface MenuUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    menuName?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    menuType?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    menuPath?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    apiEndpoint?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    iconName?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuUpdateRequest
     */
    upperMenuNo?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuUpdateRequest
     */
    menuOrder?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    menuDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    isVisible?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuUpdateRequest
     */
    isActive?: string;
}

/**
 * Check if a given object implements the MenuUpdateRequest interface.
 */
export function instanceOfMenuUpdateRequest(value: object): value is MenuUpdateRequest {
    return true;
}

export function MenuUpdateRequestFromJSON(json: any): MenuUpdateRequest {
    return MenuUpdateRequestFromJSONTyped(json, false);
}

export function MenuUpdateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuUpdateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'menuName': json['menuName'] == null ? undefined : json['menuName'],
        'menuType': json['menuType'] == null ? undefined : json['menuType'],
        'menuPath': json['menuPath'] == null ? undefined : json['menuPath'],
        'apiEndpoint': json['apiEndpoint'] == null ? undefined : json['apiEndpoint'],
        'iconName': json['iconName'] == null ? undefined : json['iconName'],
        'upperMenuNo': json['upperMenuNo'] == null ? undefined : json['upperMenuNo'],
        'menuOrder': json['menuOrder'] == null ? undefined : json['menuOrder'],
        'menuDescription': json['menuDescription'] == null ? undefined : json['menuDescription'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
    };
}

export function MenuUpdateRequestToJSON(json: any): MenuUpdateRequest {
    return MenuUpdateRequestToJSONTyped(json, false);
}

export function MenuUpdateRequestToJSONTyped(value?: MenuUpdateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'menuName': value['menuName'],
        'menuType': value['menuType'],
        'menuPath': value['menuPath'],
        'apiEndpoint': value['apiEndpoint'],
        'iconName': value['iconName'],
        'upperMenuNo': value['upperMenuNo'],
        'menuOrder': value['menuOrder'],
        'menuDescription': value['menuDescription'],
        'isVisible': value['isVisible'],
        'isActive': value['isActive'],
    };
}