// src/api/generated/models/MenuCreateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MenuCreateRequest
 */
export interface MenuCreateRequest {
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    menuCode: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    menuName: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    menuType?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    menuPath?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    apiEndpoint?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    iconName?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuCreateRequest
     */
    upperMenuNo?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuCreateRequest
     */
    menuOrder?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    menuDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    isVisible?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuCreateRequest
     */
    isActive?: string;
}

/**
 * Check if a given object implements the MenuCreateRequest interface.
 */
export function instanceOfMenuCreateRequest(value: object): value is MenuCreateRequest {
    if (!('menuCode' in value) || value['menuCode'] === undefined) return false;
    if (!('menuName' in value) || value['menuName'] === undefined) return false;
    return true;
}

export function MenuCreateRequestFromJSON(json: any): MenuCreateRequest {
    return MenuCreateRequestFromJSONTyped(json, false);
}

export function MenuCreateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuCreateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'menuCode': json['menuCode'],
        'menuName': json['menuName'],
        'menuType': json['menuType'] == null ? undefined : json['menuType'],
        'menuPath': json['menuPath'] == null ? undefined : json['menuPath'],
        'apiEndpoint': json['apiEndpoint'] == null ? undefined : json['apiEndpoint'],
        'iconName': json['iconName'] == null ? undefined : json['iconName'],
        'upperMenuNo': json['upperMenuNo'] == null ? undefined : json['upperMenuNo'],
        'menuOrder': json['menuOrder'] == null ? undefined : json['menuOrder'],
        'menuDescription': json['menuDescription'] == null ? undefined : json['menuDescription'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'isActive': json['isActive'] == null ? undefined : json['isActive'],
    };
}

export function MenuCreateRequestToJSON(json: any): MenuCreateRequest {
    return MenuCreateRequestToJSONTyped(json, false);
}

export function MenuCreateRequestToJSONTyped(value?: MenuCreateRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'menuCode': value['menuCode'],
        'menuName': value['menuName'],
        'menuType': value['menuType'],
        'menuPath': value['menuPath'],
        'apiEndpoint': value['apiEndpoint'],
        'iconName': value['iconName'],
        'upperMenuNo': value['upperMenuNo'],
        'menuOrder': value['menuOrder'],
        'menuDescription': value['menuDescription'],
        'isVisible': value['isVisible'],
        'isActive': value['isActive'],
    };
}